pyproject.toml
[tool.poetry]
name = "ghostpost-backend"
version = "0.1.0"
description = "Ghostpost backend application"
authors = ["Howard"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
